+++ packages/app/src/systems/Send/components/Send/Send.test.tsx
@@ -1,27 +1,13 @@
 import { fireEvent, render, screen, waitFor } from '@fuel-ui/test-utils';
-// import type { Account } from '@fuel-wallet/types';
-// import { Wallet } from 'fuels';
-
-// import { sendLoader } from '../../__mocks__/send';
 
 import { Select } from './Send.stories';
 
 import { TestWrapper } from '~/systems/Core';
 
 describe('Send', () => {
-  // let loaded: any;
-  // let acc: Account | undefined;
-  // const wallet = Wallet.generate();
-  // const loader = sendLoader(wallet);
-
   function findAssetSelect() {
     return screen.findByLabelText('Select Asset');
   }
-
-  // beforeEach(async () => {
-  // loaded = await loader();
-  // acc = loaded.acc1;
-  // });
 
   describe('Send.Select()', () => {
     it('should have all inputs rendered by default', async () => {
@@ -48,19 +34,4 @@ describe('Send', () => {
       expect(await screen.findByText('Fee (network)')).toBeInTheDocument();
     });
   });
-
-  /* describe('Send.Confirm()', () => {
-    it('should show TxFromTo component', async () => {
-      const Content = await storyToComponent(Confirm, {}, loaded);
-      render(<Content />, { wrapper: TestWrapper });
-      await waitFor(() => screen.findByText('Ethereum'));
-
-      const addr1B256 = Address.fromPublicKey(acc!.publicKey).toB256();
-      const addr1 = shortAddress(addr1B256);
-      const addr2 = shortAddress(wallet.address.toB256());
-
-      expect(await screen.findByText(addr1)).toBeInTheDocument();
-      expect(await screen.findByText(addr2)).toBeInTheDocument();
-    });
-  }); */
 });
